+++ src/importers/baseImporter.ts
@@ -14,7 +14,7 @@ import { LoginView } from '../models/view/loginView';
 import { SecureNoteView } from '../models/view/secureNoteView';
 
 import { LogService } from '../abstractions/log.service';
-import { ConsoleLogService } from '../cli/services/consoleLog.service';
+import { ConsoleLogService } from '../services/consoleLog.service';
 import { CipherType } from '../enums/cipherType';
 import { FieldType } from '../enums/fieldType';
 import { SecureNoteType } from '../enums/secureNoteType';
